seperated admindata into 2 files
