+++ social-apple/Classes/AdminStateData.swift
@@ -1,5 +1,5 @@
 //
-//  AdminData.swift
+//  AdminStateData.swift
 //  social-apple
 //
 //  Created by Daniel Kravec on 2024-04-11.
@@ -27,7 +27,7 @@ class AdminErrorFeed: ObservableObject {
             }
             self.client.api.admin.errors.list() { result in
                 print("allpost request")
-                
+
                 switch result {
                 case .success(let feed):
                     DispatchQueue.main.async {
@@ -46,7 +46,7 @@ class AdminErrorFeed: ObservableObject {
             }
         }
     }
-    
+
     func addIssues(newIssues: [ErrorIssueData], toClear:Bool=false) -> Void{
         var addIssues = newIssues
         addIssues.reverse()
@@ -57,7 +57,7 @@ class AdminErrorFeed: ObservableObject {
             if (toClear==true) {
                 self.issues = []
             }
-            
+
             for newIssue in addIssues {
                 if let existingIndex = self.issues.firstIndex(where: { $0._id == newIssue._id }) {
                     print("existing")
@@ -68,12 +68,12 @@ class AdminErrorFeed: ObservableObject {
             }
         }
     }
-    
+
     func refreshFeed() -> Void {
         DispatchQueue.main.async {
             self.client.api.admin.errors.list() { result in
                 self.client.hapticPress()
-                
+
                 switch result {
                 case .success(let feedData):
                     DispatchQueue.main.async {
@@ -86,12 +86,12 @@ class AdminErrorFeed: ObservableObject {
             }
         }
     }
-    
+
     func nextIndex() -> Void {
         DispatchQueue.main.async {
             self.client.api.admin.errors.list(indexID: self.errorIndex.prevIndexID ?? "") { result in
                 self.client.hapticPress()
-                
+
                 switch result {
                 case .success(let feed):
                     DispatchQueue.main.async {
@@ -105,57 +105,4 @@ class AdminErrorFeed: ObservableObject {
             }
         }
     }
-}
-
-struct ErrorIndexData: Decodable {
-    var indexID: String
-    var nextIndexID: String? = nil
-    var prevIndexID: String? = nil
-    var amount: Int64? = 0
-    var timestamp: Int64
-    var foundIssues: [ErrorIssueData]
-}
-
-struct ErrorIssueData: Decodable, Identifiable, Equatable {
-    static func == (lhs: ErrorIssueData, rhs: ErrorIssueData) -> Bool {
-        return (lhs._id == rhs._id)
-    }
-    
-    var id = UUID()
-    var _id: String
-    var userID: String? = ""
-    var errorVersion: Int64
-    var errorCode: String
-    var errorMsg: String
-    var timestamp: Int64
-    var resolved: Bool
-    var resolvedTimestamp: Int64? = nil
-    var inReview: Bool
-    var reviewedBy: String? = nil
-    var reviewTimestamp: Int64? = nil
-    var reviewHistory: [ErrorIssueHistory]? = []
-    
-    private enum CodingKeys: String, CodingKey {
-        case _id
-        case userID
-        case errorVersion
-        case errorCode
-        case errorMsg
-        case timestamp
-        case resolved
-        case resolvedTimestamp
-        case inReview
-        case reviewedBy
-        case reviewTimestamp
-        case reviewHistory
-    }
-
-}
-
-struct ErrorIssueHistory: Decodable {
-    var _id: String
-    var reviewBy: String
-    var reviewStart: Int64
-    var reviewEnd: Int64
-    var resolvedTimestamp: Int64
 }
